Make BigQuery setupPlugin cheaper
`setupPlugin` is called whenever _any_ thread on any plugin server boots
or a plugin is enabled. These were failing regularly due to socket
hangups.

On the other hand, this export was using setupPlugin to do BigQuery
setup work over and over again. This change makes it so we do that setup
only once when needed